refactor: 음성 녹음 시작 및 정지 로직 분리
